A test dataset

A test sample data used for testing analysis functions. It differs from reviews data. The text column in this data frame is similar to the comment column in the reviews data frame. Note that this data frame contains 170 texts that are in common, verbatim, with comments in the reviews dataset. This is because some users made the same comments. The id column shows that these are not the same observations, just similarly worded comments from different reviews.

Usage

reviews_test

Format

A tbl_df with with 1,481 rows and 4 variables:

rating

Rating score, out of 5.

text

Comment text, in Turkish.

emotion

n for negative, p for positive.

id

Rating ID.

Examples

reviews_test
#> # A tibble: 1,481 × 4
#>    rating text                                                     emotion    id
#>     <dbl> <chr>                                                    <chr>   <int>
#>  1      5 Kesinlikle alın aldırın pahalı markalara o kadar para v… p           1
#>  2      5 10 yıldız olsa veririm super                             p           2
#>  3      5 ürün 5 yıldız hakkediyor, ayrıca hem m hem de l beden a… p           3
#>  4      1 Yorumlara bakarak aldım çok kalitesiz hiç güzel değil..… n           4
#>  5      5 1 75 70 kilo 38 beden aldım 40 da olurmuş üstten oturdu… p           5
#>  6      5 10 yıldız olsa 10 verirdim harika gerçekten . Basenliyi… p           6
#>  7      5 On numara beş yıldız bir beden küçük söyleyin normalde … p           7
#>  8      5 Pantolon çok güzel geldi rikralı çok esnek ve çok rahat… p           8
#>  9      5 Tam bedeninizi alın kesinlikle çok sık duruyo ustte ben… p           9
#> 10      5 ürün kesimi kumaşı çok güzel gayet esnek ve kaliteli üs… p          10
#> # ℹ 1,471 more rows